[optimize] bytes instead of ints for local iteration indices

+++ lib/euclidean/euclidean.c
@@ -72,15 +72,15 @@ uint16_t euclid(uint8_t length, uint8_t density, uint8_t offset) {
   if (per_pulse > 0 && remainder < 2) { 
     // Handle easy case, when there is a 0 or 1 remainder
 
-    for (int a = 0; a < density; a++) {
-      for (int b = workbeat_count - 1; b > workbeat_count - per_pulse - 1; b--) {
+    for (uint8_t a = 0; a < density; a++) {
+      for (uint8_t b = workbeat_count - 1; b > workbeat_count - per_pulse - 1; b--) {
         workbeat[a] = binary_concat(workbeat[a], workbeat[b]);
       }
       workbeat_count = workbeat_count - per_pulse;
     }
 
     outbeat = 0; // Concatenate workbeat into outbeat - according to workbeat_count
-    for (int a = 0; a < workbeat_count; a++) {
+    for (uint8_t a = 0; a < workbeat_count; a++) {
       outbeat = binary_concat(outbeat, workbeat[a]);
     }
